demo_image_search/requirements.txt
gradio
spaces
sentence-transformers
faiss-gpu-cu12
pillow
numpy
torch
torchvision